# Performance Improvements for V1 calculation