--- a/go.mod
+++ b/go.mod
@@ -4,14 +4,12 @@ go 1.18
 
 require (
 	github.com/google/gopacket v1.1.19
-	github.com/samber/lo v1.38.1
 	github.com/stretchr/testify v1.8.4
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/exp v0.0.0-20230817173708-d852ddb80c63 // indirect
 	golang.org/x/net v0.15.0 // indirect
 	golang.org/x/sys v0.12.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect