Hello,

This PR continues on from https://github.com/satta/gommunityid/pull/6. I've found a way to further optimise the `Hash` function without extra added complexity, as well as removing the dependency on `lo`.

Bellow is the new benchstat, overall, this appears to reduce the allocations from 3 to 2, as well as reduce the time taken by another 14.38%. 

```
goos: darwin
goarch: arm64
pkg: github.com/satta/gommunityid
             │ master.txt  │           this_branch.txt           │
             │   sec/op    │   sec/op     vs base                │
Calc/IPv4-10   155.0n ± 0%   131.6n ± 0%  -15.09% (p=0.000 n=30)
Calc/IPv6-10   144.9n ± 1%   125.1n ± 0%  -13.67% (p=0.000 n=30)
geomean        149.9n        128.3n       -14.38%

             │ master.txt │          this_branch.txt           │
             │    B/op    │    B/op     vs base                │
Calc/IPv4-10   152.0 ± 0%   136.0 ± 0%  -10.53% (p=0.000 n=30)
Calc/IPv6-10   184.0 ± 0%   136.0 ± 0%  -26.09% (p=0.000 n=30)
geomean        167.2        136.0       -18.68%

             │ master.txt │          this_branch.txt           │
             │ allocs/op  │ allocs/op   vs base                │
Calc/IPv4-10   3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=30)
Calc/IPv6-10   3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=30)
geomean        3.000        2.000       -33.33%
```

If approved, would it be possible to get a minor release of this package done, to make using the performance changes in other go packages easier?